separtes api and db
