+++ services/web/server/src/simcore_service_webserver/payments/api.py
@@ -1,14 +1,16 @@
 from ._api import (
-    cancel_creation_of_wallet_payment_method,
     cancel_payment_to_wallet,
     create_payment_to_wallet,
-    delete_wallet_payment_method,
     get_user_payments_page,
+)
+from ._client import get_payments_service_api
+from ._methods_api import (
+    cancel_creation_of_wallet_payment_method,
+    delete_wallet_payment_method,
     get_wallet_payment_method,
     init_creation_of_wallet_payment_method,
     list_wallet_payment_methods,
 )
-from ._client import get_payments_service_api
 
 __all__: tuple[str, ...] = (
     "cancel_payment_to_wallet",
